chore: file restructuring and formatting
